Starting initial restructuring
The old code base was a bit messy. Both Zetta and I have learned a lot since we started that code. Some of it we knew while we were making it that it was spaghetti but we weren't exactly in a position at that time to refactor. Now begins the process.